+++ doghouse/client/github_client.go
@@ -3,7 +3,7 @@ package client
 import (
 	"context"
 
-	"github.com/google/go-github/v53/github"
+	"github.com/google/go-github/v60/github"
 
 	"github.com/reviewdog/reviewdog/doghouse"
 	"github.com/reviewdog/reviewdog/doghouse/server"
@@ -15,6 +15,6 @@ type GitHubClient struct {
 	Client *github.Client
 }
 
-func (c *GitHubClient) Check(ctx context.Context, req *doghouse.CheckRequest) (*doghouse.CheckResponse, error) {
-	return server.NewChecker(req, c.Client).Check(ctx)
+func (c *GitHubClient) Check(ctx context.Context, req *doghouse.CheckRequest, makeRequest bool) (*doghouse.CheckResponse, error) {
+	return server.NewChecker(req, c.Client).Check(ctx, makeRequest)
 }