+++ CHANGELOG.md
@@ -17,12 +17,40 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### :rotating_light: Breaking changes
 - ...
 
----
+## [v0.17.2] - 2024-03-11
 
 ### :bug: Fixes
-- [#967](https://github.com/reviewdog/reviewdog/pull/967) Fix parsing long lines in diffs #967
-- [#1426](https://github.com/reviewdog/reviewdog/pull/1426) Remove default error level
+- [#933](https://github.com/reviewdog/reviewdog/issues/933) Fix go vet errorformat
 
+## [v0.17.1] - 2024-02-08
+
+### :bug: Fixes
+- [#1651](https://github.com/reviewdog/reviewdog/pull/1651) Revert #1576: Support `--filter-mode=file` in `github-pr-review`. Reasons: [#1645](https://github.com/reviewdog/reviewdog/pull/1645)
+- [#1653](https://github.com/reviewdog/reviewdog/pull/1653) fix: SARIF parser: parse with no region result. fix originalOutput field
+- [#1657](https://github.com/reviewdog/reviewdog/pull/1657) Fix sending incorrect line numbers to BitBucket Server Code Insight API. (fixes #1652)
+
+## [v0.17.0] - 2024-01-22
+
+### :rocket: Enhancements
+- [#1623](https://github.com/reviewdog/reviewdog/pull/1623) Add reporter for GitHub PR annotations `github-pr-annotations`
+
+## [v0.16.0] - 2023-12-17
+
+### :rocket: Enhancements
+- [#1573](https://github.com/reviewdog/reviewdog/pull/1573) Add filter tests for file/nofilter mode
+- [#1576](https://github.com/reviewdog/reviewdog/pull/1576) Support `--filter-mode=file` in `github-pr-review`
+- [#1596](https://github.com/reviewdog/reviewdog/pull/1596) Use `CI_MERGE_REQUEST_DIFF_BASE_SHA` envvar if available in `gitlab-mr-discussion`
+- [#1521](https://github.com/reviewdog/reviewdog/pull/1521) strict check of pr-review write permission
+- [#1617](https://github.com/reviewdog/reviewdog/pull/1617) Add reporter to Gitea PR review comments `gitea-pr-review`
+
+---
+
+## [v0.15.0] - 2023-09-02
+
+### :rocket: Enhancements
+- [#1554](https://github.com/reviewdog/reviewdog/pull/1554) Add SARIF format input support.
+
+---
 
 ## [v0.14.2] - 2023-06-17
 
@@ -156,7 +184,7 @@ $ cd subdir/ && reviewdog -filter-mode=file -fail-on-error -reporter=github-pr-r
 
 See https://github.com/reviewdog/reviewdog/releases for older release note.
 
-[Unreleased]: https://github.com/reviewdog/reviewdog/compare/v0.14.2...HEAD
+[Unreleased]: https://github.com/reviewdog/reviewdog/compare/v0.17.1...HEAD
 [v0.10.0]: https://github.com/reviewdog/reviewdog/compare/v0.9.17...v0.10.0
 [v0.10.1]: https://github.com/reviewdog/reviewdog/compare/v0.10.0...v0.10.1
 [v0.10.2]: https://github.com/reviewdog/reviewdog/compare/v0.10.1...v0.10.2
@@ -167,4 +195,9 @@ See https://github.com/reviewdog/reviewdog/releases for older release note.
 [v0.14.0]: https://github.com/reviewdog/reviewdog/compare/v0.13.1...v0.14.0
 [v0.14.1]: https://github.com/reviewdog/reviewdog/compare/v0.14.0...v0.14.1
 [v0.14.2]: https://github.com/reviewdog/reviewdog/compare/v0.14.1...v0.14.2
+[v0.15.0]: https://github.com/reviewdog/reviewdog/compare/v0.14.2...v0.15.0
+[v0.16.0]: https://github.com/reviewdog/reviewdog/compare/v0.15.0...v0.16.0
+[v0.17.0]: https://github.com/reviewdog/reviewdog/compare/v0.16.0...v0.17.0
+[v0.17.1]: https://github.com/reviewdog/reviewdog/compare/v0.17.0...v0.17.1
+[v0.17.2]: https://github.com/reviewdog/reviewdog/compare/v0.17.1...v0.17.2
 [@haya14busa]: https://github.com/haya14busa